fix: slashcommand builder type split (#10253)
