+++ packages/builders/src/interactions/slashCommands/SlashCommandBuilder.ts
@@ -1,13 +1,14 @@
 import type { APIApplicationCommandOption, LocalizationMap, Permissions } from 'discord-api-types/v10';
 import { mix } from 'ts-mixer';
 import { SharedNameAndDescription } from './mixins/NameAndDescription.js';
+import { SharedSlashCommand } from './mixins/SharedSlashCommand.js';
 import { SharedSlashCommandOptions } from './mixins/SharedSlashCommandOptions.js';
 import { SharedSlashCommandSubcommands } from './mixins/SharedSubcommands.js';
 
 /**
  * A builder that creates API-compatible JSON data for slash commands.
  */
-@mix(SharedSlashCommandOptions, SharedNameAndDescription, SharedSlashCommandSubcommands)
+@mix(SharedSlashCommandOptions, SharedNameAndDescription, SharedSlashCommandSubcommands, SharedSlashCommand)
 export class SlashCommandBuilder {
 	/**
 	 * The name of this command.
@@ -37,7 +38,7 @@ export class SlashCommandBuilder {
 	/**
 	 * Whether this command is enabled by default when the application is added to a guild.
 	 *
-	 * @deprecated Use {@link SharedSlashCommandSubcommands.setDefaultMemberPermissions} or {@link SharedSlashCommandSubcommands.setDMPermission} instead.
+	 * @deprecated Use {@link SharedSlashCommand.setDefaultMemberPermissions} or {@link SharedSlashCommand.setDMPermission} instead.
 	 */
 	public readonly default_permission: boolean | undefined = undefined;
 
@@ -63,14 +64,16 @@ export class SlashCommandBuilder {
 export interface SlashCommandBuilder
 	extends SharedNameAndDescription,
 		SharedSlashCommandOptions<SlashCommandOptionsOnlyBuilder>,
-		SharedSlashCommandSubcommands<SlashCommandSubcommandsOnlyBuilder> {}
+		SharedSlashCommandSubcommands<SlashCommandSubcommandsOnlyBuilder>,
+		SharedSlashCommand {}
 
 /**
  * An interface specifically for slash command subcommands.
  */
 export interface SlashCommandSubcommandsOnlyBuilder
 	extends SharedNameAndDescription,
-		SharedSlashCommandSubcommands<SlashCommandSubcommandsOnlyBuilder> {}
+		SharedSlashCommandSubcommands<SlashCommandSubcommandsOnlyBuilder>,
+		SharedSlashCommand {}
 
 /**
  * An interface specifically for slash command options.
@@ -78,7 +81,7 @@ export interface SlashCommandSubcommandsOnlyBuilder
 export interface SlashCommandOptionsOnlyBuilder
 	extends SharedNameAndDescription,
 		SharedSlashCommandOptions<SlashCommandOptionsOnlyBuilder>,
-		ToAPIApplicationCommandOptions {}
+		SharedSlashCommand {}
 
 /**
  * An interface that ensures the `toJSON()` call will return something
